Added non-mechanical feats
In anticipation of mechanical feats (i.e. those that change statistics
of the character), a trait Feat has been defined and both mechanical
and non-mechanical feats will implement it. After fighting against the
borrow checker for a while, it became evident that Rc<Trait> is much
better than Box<Trait> for capturing Traits into the store of the
builder since Rc<_> automatically lives as long as the surrounding
struct. Directly transferring ownership of an "impl Trait" is not
possible because it is possible to implement Traits on references and
other objects with specific lifetimes.